refactor(orchestrator): organize curriculum package
