+++ examples/advanced/intellect-3.1/rl.toml
@@ -50,7 +50,7 @@ oversampling_factor = 2
 [[orchestrator.train.source]]
 name = "swe"
 ratio = 0.3
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "r2e-gym"
@@ -65,7 +65,7 @@ labels = ["intellect-3.1", "swe"]
 [[orchestrator.train.source]]
 name = "deepdive"
 ratio = 0.2
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "deepdive"
@@ -82,7 +82,7 @@ labels = ["intellect-3.1", "deepdive"]
 [[orchestrator.train.source]]
 name = "math"
 ratio = 0.3
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "i3_math"
@@ -96,7 +96,7 @@ type = "subprocess"
 [[orchestrator.train.source]]
 name = "logic"
 ratio = 0.2
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "i3_logic"
@@ -113,7 +113,7 @@ type = "subprocess"
 [[orchestrator.train.source]]
 name = "code"
 ratio = 0.2
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "i3_code"
