+++ examples/basic/wiki-search/rl.toml
@@ -42,7 +42,7 @@ max_completion_tokens = 512
 
 [[orchestrator.train.source]]
 name = "wiki-search"
-curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curricula.AdvantageRangeGate" } } }
+curriculum = { gates = { zero_advantage = { import_path = "prime_rl.orchestrator.curriculum.AdvantageRangeGate" } } }
 
 [orchestrator.train.source.env.taskset]
 id = "wiki-search"
